grouped lot of code into function
